// src/api/generated/models/ServerStatsResponseDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */

export type ServerStatsResponseDTO = {
    onlineCount: number;
    totalCount: number;
    bannedUserCount: number;
    serverId: string;
    serverName: string;
    ownerName: string;
    createdAt: string;
    allTimeHigh: number;
    newestMember: string;
    channelCount: number;
    emojiCount: number;
};